requirements.txt
anyio==4.0.0
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
flake8==6.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
isort==5.12.0
lxml==4.9.3
mccabe==0.7.0
outcome==1.2.0
pycodestyle==2.11.0
pyflakes==3.1.0
PySocks==1.7.1
python-dotenv==1.0.0
python-telegram-bot==20.5
requests==2.31.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.22.2
trio-websocket==0.10.4
urllib3==2.0.4
wsproto==1.2.0
